Написал логику MenuScreen.kt

+++ app/src/main/java/com/pivnoydevelopment/cafeapp/features/menu/data/dto/MenuItemDto.kt
@@ -4,5 +4,5 @@ data class MenuItemDto(
     val id: Int,
     val name: String,
     val imageURL: String,
-    val price: Double
+    val price: Int
 )
